音楽、動画ファイルのPodcast RSS feedファイルを探してindex.htmlにリンクを張る処理の実装

+++ app/startup.sh
@@ -3,7 +3,7 @@
 set -eu
 
 cd /usr/local/apache2/htdocs/
-if [ ! -e "channels" ]; then
+if [ ! -e "channels" ] && [ ! -L "channels" ]; then
     ln -s /volumes/channels channels
 fi
 # if [ ! -e "thumbs" ]; then
